Add files via upload
Affects the system to load skins associated with a specific file slot. Fixes the bug of CustomKnight not working with MoreSaves in files past the default 4. Stores skins associated with each slot as a Dictionary rather than an Array, with the slot number (starting from 0 to be consistent with existing array calls) being the key value and returning the relevant skin. If a save slot is loaded that does not currently have an entry in the Dictionary, it will be given the appropriate entry.